feat: ディレクトリ構造の整理
