+++ src/app/contact/page.tsx
@@ -1,4 +1,4 @@
-import { Contact } from "../_components/Contact";
+import { Contact } from "./_components/Contact";
 
 export default function ContactPage() {
   return <Contact />;
