--- src/app/data/posts.ts
@@ -1,8 +1,0 @@
-export interface Post {
-  id: string;
-  title: string;
-  content: string;
-  thumbnailUrl: string;
-  createdAt: string;
-  categories: string[];
-}
