+++ src/app/globals.css
@@ -13,6 +13,7 @@
     --foreground: #ededed;
   }
 }
+/* hennkou */
 
 body {
   color: var(--foreground);
