+++ src/app/page.tsx
@@ -1,5 +1,4 @@
-import { Home } from "./_components/Home";
-import Image from "next/image";
+import { Home } from './_components/Home';
 
 export default function Page() {
   return <Home />;
